Added in score graphics
So far I can't get the sprite to fade going up.. not sure why.   It only
gets triggered right now on a direct hit to a cat.   We still have bugs
to figure out with the cat deaths.. it seems TOO hard to kill them.  If
they fall to the ground they should probably die.